Refs #3026:  Added a field for Wiki description to Algorithm. Made WikiMaker use it to generate wiki pages.

+++ Code/Mantid/Framework/Crystal/test/MaskPeaksWorkspaceTest.h
@@ -46,7 +46,7 @@ public:
       in_ws *= 2.0;
     if (type == WEIGHTED_NOTIME)
     {
-      for (int i =0; i<in_ws->getNumberHistograms(); i++)
+      for (size_t i =0; i<in_ws->getNumberHistograms(); i++)
       {
         EventList & el = in_ws->getEventList(i);
         el.compressEvents(0.0, &el);
